requirements.txt
langgraph
google-generativeai
python-dotenv
pillow
